+++ src/main.ts
@@ -1,1 +1,2 @@
 export * from "./datetime";
+export type { IDuration, DateTimeRange } from "./duration";